requirements.txt
boto3
fastapi
botocore
python-dotenv
python-multipart
uvicorn